backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
pyshark==0.6.0
pycryptodome==3.19.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.2
gunicorn==21.2.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 